+++ types/prompt.ts
@@ -1,11 +1,9 @@
-import {OpenAIModelID} from "@/types/openai"
-
+// This type is used to store a prompt in the store.
 export interface Prompt {
   id: string
   name: string
   description: string
   content: string
-  modelId: OpenAIModelID
   folderId: string | undefined
   factory: boolean | undefined
 }